+++ src/ts/EventHandlers/VideoEventHandlers.ts
@@ -7,8 +7,6 @@ import { UnityAdsError } from 'Constants/UnityAdsError';
 import { Diagnostics } from 'Utilities/Diagnostics';
 import { DiagnosticError } from 'Errors/DiagnosticError';
 import { VideoAdUnit } from 'AdUnits/VideoAdUnit';
-import { PerformanceCampaign } from 'Models/PerformanceCampaign';
-import { VastCampaign } from 'Models/Vast/VastCampaign';
 import { TestEnvironment } from 'Utilities/TestEnvironment';
 import { ViewConfiguration } from 'AdUnits/Containers/AdUnitContainer';
 
@@ -22,18 +20,8 @@ export class VideoEventHandlers {
 
         if(duration > 40000) {
             const campaign = adUnit.getCampaign();
-            let url: string;
-            let originalUrl: string;
-            if(campaign instanceof PerformanceCampaign) {
-                url = (<PerformanceCampaign>campaign).getVideo().getUrl();
-                originalUrl = (<PerformanceCampaign>campaign).getVideo().getOriginalUrl();
-            } else if(campaign instanceof VastCampaign) {
-                url = (<VastCampaign>campaign).getVideo().getUrl();
-                originalUrl = (<VastCampaign>campaign).getVideo().getOriginalUrl();
-            } else {
-                throw new Error('Unknown campaign type');
-            }
-
+            const url = adUnit.getVideo().getUrl();
+            const originalUrl = adUnit.getVideo().getOriginalUrl();
             const error: DiagnosticError = new DiagnosticError(new Error('Too long video'), {
                 duration: duration,
                 campaignId: campaign.getId(),
@@ -180,7 +168,7 @@ export class VideoEventHandlers {
     }
 
     public static onVideoCompleted(sessionManager: SessionManager, adUnit: VideoAdUnit): void {
-        adUnit.getVideo().setActive(false);
+        adUnit.setActive(false);
         adUnit.setFinishState(FinishState.COMPLETED);
         sessionManager.sendView(adUnit);
 
@@ -282,7 +270,7 @@ export class VideoEventHandlers {
 
     private static handleVideoError(nativeBridge: NativeBridge, videoAdUnit: VideoAdUnit) {
         videoAdUnit.getVideo().setErrorStatus(true);
-        videoAdUnit.getVideo().setActive(false);
+        videoAdUnit.setActive(false);
         videoAdUnit.setFinishState(FinishState.ERROR);
 
         this.updateViewsOnVideoError(videoAdUnit);