+++ src/ts/Test/Unit/TestHelpers/TestFixtures.ts
@@ -13,6 +13,7 @@ import { PerformanceCampaign } from 'Models/PerformanceCampaign';
 import { MRAIDCampaign } from 'Models/MRAIDCampaign';
 import { Configuration } from 'Models/Configuration';
 import { PromoCampaign } from 'Models/PromoCampaign';
+import { ICacheDiagnostics } from 'Utilities/Cache';
 
 export class TestFixtures {
 
@@ -115,5 +116,14 @@ export class TestFixtures {
             gamerId: 'abc123',
             abGroup: 0
         });
+    }
+
+    public static getCacheDiagnostics(): ICacheDiagnostics {
+        return {
+            creativeType: 'TEST',
+            gamerId: '1234abcd',
+            targetGameId: 5678,
+            targetCampaignId: '123456abcdef'
+        };
     }
 }